43/48p 382/396p 0f 243/248p

+++ ft_intlen.c
@@ -6,13 +6,13 @@
 /*   By: sleung <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/11 13:35:02 by sleung            #+#    #+#             */
-/*   Updated: 2017/03/06 13:40:20 by sleung           ###   ########.fr       */
+/*   Updated: 2017/03/09 18:07:43 by sleung           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-int	ft_intlen(int nb)
+int	ft_intlen(intmax_t nb)
 {
 	int	c;
