+++ handle_sharp.c
@@ -6,13 +6,13 @@
 /*   By: sleung <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/21 11:11:25 by sleung            #+#    #+#             */
-/*   Updated: 2017/03/06 13:41:36 by sleung           ###   ########.fr       */
+/*   Updated: 2017/03/09 16:10:52 by sleung           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-char	*handle_sharp(t_struct *d, char **tmp, int *ti, int n)
+char	*handle_sharp(t_struct *d, char **tmp, int *ti, intmax_t n)
 {
 	char	*ctmp;
 	int		j;
